// src/api/generated/schemas/ongoingFestivalsDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * TripBite API
 * OpenAPI spec version: 1.0
 */
import type { OngoingFestivalItemDto } from './ongoingFestivalItemDto';
import type { OngoingFestivalsDtoType } from './ongoingFestivalsDtoType';

export interface OngoingFestivalsDto {
  type: OngoingFestivalsDtoType;
  items: OngoingFestivalItemDto[];
}